Extract help.py as a separate file
